+++ src/main/java/api/soldout/io/soldout/service/trade/TradeService.java
@@ -1,7 +1,6 @@
 package api.soldout.io.soldout.service.trade;
 
 import api.soldout.io.soldout.dtos.entity.TradeDto;
-import api.soldout.io.soldout.util.enums.SaleType;
 import java.util.List;
 
 /**
@@ -11,8 +10,6 @@ import java.util.List;
 public interface TradeService {
 
   void saveTrade(int productId, int orderId, int saleId, int size, int price);
-
-  void matchTradeByOrder(int orderId, int productId, int size, int price);
 
   List<TradeDto> findByOrderId(int orderId);
